Refactor ingredient quantity type: change quantity from float to int in Ingredient struct and update related supply chain message handling

+++ include/products.h
@@ -50,7 +50,7 @@ typedef enum {
 // Struct for an ingredient with quantity
 typedef struct {
     IngredientType type; // Type of ingredient
-    float quantity;
+    int quantity;
 } Ingredient;
 
 // Updated Product struct with separate timing fields
